+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Restructure the definition of the Region ID constants. !1325 @theFeiter
 - Remove moment.js dependency. !1303 #678 @ctwx_ok
 - Moved the button for new stores to vue store list !1282 !1339 @chriswalg
+- Refactored wakeupSleepingUsers to MaintenanceGateway !1301 @Caluera
 
 ## Dev/Test/CI stuff
 - Add "linux" tag for finding CI servers !1332 @nicksellen